Refactor game logic and clean up scenes.py
Moved the game loop and input handling from scenes.py to a new main.py file, improving code organization. Removed commented-out code and unnecessary functions from scenes.py. Deleted the unused todo.txt file.
